Add bosh managed SG to ELBs
This is not intended to be a long term solution, but to avoid creating
SGs for the ELB destinations and assign them to the job destination.
All the members of the bosh_managed SG can talk to each other freely,
so ELBs can talk to all memebers (and members can talk to ELBs).