Add collectd to deployer concourse
Add 0.4 release of our paas-collectd-boshrelease collectd to concourse.
This version contains fixes required to be able to run on concourse.

The initial reason why we used fork was that we could add custom tags and
pin our deployment to these tags.
Now, because bosh-init can only consume tarballs (you can't upload a release
to bosh-init like you can to BOSH), I had to create a dev-release tarbal
(using standard bosh-cli way bosh create release --with-tarball --force
--version 0.4, more documentation about bosh releases:
https://bosh.io/docs/create-release.html#dev-release-release). I have
created the tarballs from the same code as is tagged. That is, 0.4 release
is created from code that's tagged as 0.4.

It seemed to me that a good place to put the release tarball was in github
release. This practice is also used in some CF repositories (e.g. diego). We
have a spike to analyse our situation with releases completely, but we did
not want to block this story on it (#115142265 - see comments in the story).
In the future, we can easily move the release location to elsewhere and we
can change the way to build the releases.