Refactor collectd properties in the manifest
Refactoring ready for the addition of collectd to Concourse and
Bosh[1]. The collectd configuration properties will be required
in the CloudFoundry, Concourse, and Bosh manifests, which are
all rendered seperately, and therefore should be a shared file
to avoid duplication.

The environment specific YAML file in the pipeline[2] has been moved
to preserve its precedence.

[1] https://www.pivotaltracker.com/story/show/122189093
[2] concourse/pipelines/create-bosh-cloudfoundry.yml:1114