Refactor collectd properties in the runtime config
Refactoring ready for the addition of collectd to Concourse and
Bosh[1]. The collectd configuration properties will be required
in the CloudFoundry, Concourse, and Bosh manifests, which are
all rendered seperately, and therefore should be a shared file
to avoid duplication.

The runtime config will now have to be rendered in the pipeline,
so its name has been changed to make it more specific and clear
what it is for. The environment specific YAML file in the pipeline[2]
has been moved to preserve its precedence.

[1] https://www.pivotaltracker.com/story/show/122189093
[2] concourse/pipelines/create-bosh-cloudfoundry.yml:1122