Update retention period of existing instances
We want to be sure that all the existing RDSBroker instances created
will be updated with the new retention period.

For that we add a tasks in the post-deploy job that:
 - Retrieve the retention period of the first plan in RDS Broker.
   We assume that all plans have the same retention.
 - Use AWS Cli to find all the rdsbroker-* instances for the current
   environment and update the retention period.

The change in the backup retention period is done with the option
--no-apply-immediately, so the database will updated during the
maintenance window[1]

Note that the DB will be restarted if the retention is changed from 0
to any other value.

This update operation is idempotent and can be executed several times.
If the instance already has the desired value, no action will be taken.
If there was already a change in the retention period pending for the
next maintenance window, AWS will use the latest value provided by this
script. If the new value is already the one active, AWS will not
cancel the change pending.

[1] http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/Overview.DBInstance.Modifying.html