Remove the paas_demo quota definition.
This is identical to the newly created 'small' quota, so this can be
removed, and the 'small' quota used instead.

Note: removing it from the manifest here won't cause the quota to be
deleted from CF (the quota setting script only does create/update). This
will have to be done manually once the org has been updated to use the
'small' quota.