Use awscli version with get-caller-identity
We need to use the `aws sts get-caller-identity`, required to
retrieve the account ID to generate the ARN of the RDS instances
in the script that updates the RDS instances. This is the only command
that can work on deployer with the current policy.

But this command is only available in versions > 1.10.18.

We update the repository to the current latest version